refactor(cli): split node fallback entrypoint paths
