+++ .sampo/changesets/920-split-node-fallback-entrypoint.md
@@ -1,0 +1,5 @@
+---
+npm/wp-typia: patch
+---
+
+Split Node fallback version, templates, doctor, and entrypoint error handling into focused modules.
