+++ packages/wp-typia/src/node-fallback/templates.ts
@@ -1,0 +1,97 @@
+import {
+  CLI_DIAGNOSTIC_CODES,
+  createCliCommandError,
+} from '@wp-typia/project-tools/cli-diagnostics';
+import {
+  getTemplateById,
+  listTemplates,
+} from '@wp-typia/project-tools/cli-templates';
+import { executeTemplatesCommand } from '../runtime-bridge';
+import type { NodeFallbackDispatchContext, NodeFallbackGlobalFlags } from './types';
+
+function renderNodeFallbackTemplatesJson(
+  printLine: NodeFallbackDispatchContext['printLine'],
+  flags: NodeFallbackGlobalFlags,
+  subcommand: string,
+) {
+  if (subcommand === 'list') {
+    printLine(
+      JSON.stringify(
+        {
+          templates: listTemplates(),
+        },
+        null,
+        2,
+      ),
+    );
+    return;
+  }
+
+  const templateId = flags.id;
+  if (!templateId) {
+    throw createCliCommandError({
+      code: CLI_DIAGNOSTIC_CODES.MISSING_ARGUMENT,
+      command: 'templates',
+      detailLines: ['`wp-typia templates inspect` requires <template-id>.'],
+    });
+  }
+  const template = getTemplateById(templateId);
+  if (!template) {
+    throw createCliCommandError({
+      code: CLI_DIAGNOSTIC_CODES.INVALID_ARGUMENT,
+      command: 'templates',
+      detailLines: [`Unknown template "${templateId}".`],
+    });
+  }
+  printLine(
+    JSON.stringify(
+      {
+        template,
+      },
+      null,
+      2,
+    ),
+  );
+}
+
+export async function dispatchNodeFallbackTemplates({
+  mergedFlags,
+  positionals,
+  printLine,
+}: NodeFallbackDispatchContext): Promise<void> {
+  const subcommand = positionals[1];
+  const templateId =
+    typeof mergedFlags.id === 'string'
+      ? mergedFlags.id
+      : (positionals[2] as string | undefined);
+  const resolvedSubcommand = templateId ? 'inspect' : (subcommand ?? 'list');
+  if (resolvedSubcommand !== 'list' && resolvedSubcommand !== 'inspect') {
+    throw createCliCommandError({
+      code: CLI_DIAGNOSTIC_CODES.INVALID_COMMAND,
+      command: 'templates',
+      detailLines: [
+        `Unknown templates subcommand "${resolvedSubcommand}". Expected list or inspect.`,
+      ],
+    });
+  }
+  if (mergedFlags.format === 'json') {
+    renderNodeFallbackTemplatesJson(
+      printLine,
+      {
+        format: mergedFlags.format as string | undefined,
+        id: templateId,
+      },
+      resolvedSubcommand,
+    );
+    return;
+  }
+  await executeTemplatesCommand(
+    {
+      flags: {
+        id: templateId,
+        subcommand: resolvedSubcommand,
+      },
+    },
+    printLine,
+  );
+}
