+++ packages/wp-typia/src/node-fallback/version.ts
@@ -1,0 +1,29 @@
+import packageJson from '../../package.json';
+import type { PrintLine } from '../print-line';
+
+export function renderNodeFallbackVersion(
+  printLine: PrintLine,
+  options: {
+    format?: string;
+  } = {},
+) {
+  if (options.format === 'json') {
+    printLine(
+      JSON.stringify(
+        {
+          ok: true,
+          data: {
+            type: 'version',
+            name: packageJson.name,
+            version: packageJson.version,
+          },
+        },
+        null,
+        2,
+      ),
+    );
+    return;
+  }
+
+  printLine(`wp-typia ${packageJson.version}`);
+}
